+++ CHANGELOG.md
@@ -4,3 +4,13 @@
 
 2017-06-21
 - setup webpack (cont)
+
+2017-06-23
+- apply webpack-dev-server for DEV env
+- separate env for DEV and PROD
+- add Procfile for heroku
+
+Todo:
+- eslint
+- redux
+-